remove setPageTransformerStyle method
