+++ app/src/main/java/com/example/zhpan/circleviewpager/transform/AccordionTransformer.java
@@ -14,33 +14,16 @@
  * limitations under the License.
  */
 
-package com.zhpan.bannerview.transform;
+package com.example.zhpan.circleviewpager.transform;
 
 import android.view.View;
 
-public class DepthPageTransformer extends BaseTransformer {
-
-	private static final float MIN_SCALE = 0.75f;
+public class AccordionTransformer extends BaseTransformer {
 
 	@Override
 	protected void onTransform(View view, float position) {
-		if (position <= 0f) {
-			view.setTranslationX(0f);
-			view.setScaleX(1f);
-			view.setScaleY(1f);
-		} else if (position <= 1f) {
-			final float scaleFactor = MIN_SCALE + (1 - MIN_SCALE) * (1 - Math.abs(position));
-			view.setAlpha(1 - position);
-			view.setPivotY(0.5f * view.getHeight());
-			view.setTranslationX(view.getWidth() * -position);
-			view.setScaleX(scaleFactor);
-			view.setScaleY(scaleFactor);
-		}
-	}
-
-	@Override
-	protected boolean isPagingEnabled() {
-		return true;
+		view.setPivotX(position < 0 ? 0 : view.getWidth());
+		view.setScaleX(position < 0 ? 1f + position : 1f - position);
 	}
 
 }
